requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=10.0.0

# Voice I/O
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13

# LLM Integration
openai>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization & Metrics
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Progress Bars
tqdm>=4.65.0
